+++ src/main/java/com/further/repository/CalendarDtRepository.java
@@ -4,6 +4,6 @@ import com.further.model.CalendarDt;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface CalendarDtRepository extends JpaRepository<CalendarDt, Integer> {
-    List<CalendarDt> findByUserId(int user_id);
+public interface CalendarDtRepository extends JpaRepository<CalendarDt, Long> {
+    List<CalendarDt> findByUserId(Long user_id);
 }
