+++ src/main/java/com/further/repository/GoalRepository.java
@@ -3,6 +3,6 @@ import com.further.model.Goal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface GoalRepository extends JpaRepository<Goal, Integer> {
-    List<Goal> findByUserId(int user_id);
+public interface GoalRepository extends JpaRepository<Goal, Long> {
+    List<Goal> findByUserId(Long user_id);
 }
