+++ src/main/java/com/further/repository/ProfileRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProfileRepository extends JpaRepository<Profile, Integer> {
-    Optional<Profile> findByUserId(int user_id);
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
+    Optional<Profile> findByUserId(Long user_id);
 }
